app/requirements.txt
# Streamlit App Requirements
# Stripped down for Streamlit Cloud deployment
# Only includes what the app pages actually need

streamlit>=1.35.0

# LangChain stack
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-tavily>=0.1.0
langchain-text-splitters>=0.3.0

# LangGraph
langgraph>=0.2.0

# HuggingFace
sentence-transformers>=2.7.0
huggingface-hub>=0.23.0

# Vector search
faiss-cpu>=1.7.4

# LLM providers
groq>=0.9.0

# PDF loading
pypdf>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
numpy>=1.26.0
tqdm>=4.66.0